// src/app/admin/stock/models/statut-peremption.enum.ts
export enum StatutPeremption {
  VALIDE = 'VALIDE',
  PROCHE_EXPIRATION = 'PROCHE_EXPIRATION',
  ALERTE_350_JOURS = 'ALERTE_350_JOURS',
  ALERTE_170_JOURS = 'ALERTE_170_JOURS',
  ALERTE_30_JOURS = 'ALERTE_30_JOURS',
  ALERTE_7_JOURS = 'ALERTE_7_JOURS',
  EXPIRE_AUJOURD_HUI = 'EXPIRE_AUJOURD_HUI',
  PERIME = 'PERIME'
}